+++ packages/react-query-next-experimental/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @tanstack/react-query-next-experimental
 
+## 5.100.14
+
+### Patch Changes
+
+- Updated dependencies [[`ed20b6d`](https://github.com/TanStack/query/commit/ed20b6d7541c908033acfcad92b0cd112930d1c3)]:
+  - @tanstack/react-query@5.100.14
+
+## 5.100.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-query@5.100.13
+
+## 5.100.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-query@5.100.12
+
 ## 5.100.11
 
 ### Patch Changes